CakePHP 4.x fixes
Shell wasn't working because of deprecations.
After making it work again, I had to add a `patchEntity()`method because beforeMarshall Events wasn't firing (It is fired before request data is converted into entities)
Note that `Shells` are deprecated since 3.6 and will be removed on 5.x in favor of `Commands`

Edited a typo in the docs, added some lines and created Installation.md since there was a link to it but the file did not exist.